fix(docs): Prerender the edit mode demo on the server
Load the Editor docs page: the live edit mode demo starts blank and
its highlighted code only pops in once the client attaches the
editor, so the surface visibly flashes after the page loads.

The demo rendered <File> with no prerendered HTML, so the server
only emitted an empty container and all highlighting happened client
side. Preload the demo file with preloadFile and hydrate <File> from
the resulting prerenderedHTML instead. The preload options mirror the
state the editor enforces in contentEditable mode (token transformer
on; gutter, line selection, and hover off) so the SSR markup matches
the post-attach render and doesn't re-highlight. The demo now ships
highlighted in the initial HTML.
