+++ apps/docs/app/(diffs)/docs/Editor/content.mdx
@@ -33,7 +33,7 @@ To enable edit mode on a `File` instance, import `Editor` from
 `@pierre/diffs/editor`. It is intentionally separate from the core bundle so you
 can lazy-load it when editing is optional.
 
-<EditorDemo />
+<EditorDemo prerenderedFile={editorDemoFile} />
 
 ### How It Works
